fix(core): made seperate file for byIdsSpecification
